makes StagedImportDataset reusable
