--- application/ui/src/features/dataset/import-export/import-jobs-list/staged-import-dataset/staged-import-dataset.component.tsx
@@ -1,71 +1,0 @@
-// Copyright (C) 2026 Intel Corporation
-// SPDX-License-Identifier: Apache-2.0
-
-import { Button, dimensionValue, Divider, Flex, Text, View } from '@geti/ui';
-import { InfoOutline } from '@geti/ui/icons';
-import { useStagedDataset } from 'hooks/api/staged-dataset.hook';
-import { useImportDatasetToProject } from 'hooks/localStorage/use-import-dataset-to-project.hook';
-
-import { DeleteStagedFileConfirmation } from '../../../../../components/delete-staged-file-confirmation/delete-staged-file-confirmation.component';
-import { formatBytes } from '../../../../../shared/util';
-import { useImportDatasetDialogState } from '../../../providers/export-import-dataset-dialog-provider.component';
-
-type StagedImportDatasetProps = {
-    fileName: string;
-    stagedDatasetId: string;
-};
-
-export const StagedImportDataset = ({ stagedDatasetId, fileName }: StagedImportDatasetProps) => {
-    const { deleteImportEntry } = useImportDatasetToProject();
-    const { error, isError, isFetching, data: stagedDataset } = useStagedDataset(stagedDatasetId);
-    const { datasetImportDialogState, setCurrentStep, setCurrentStagedId } = useImportDatasetDialogState();
-
-    const handleOpen = () => {
-        setCurrentStep('labelMapping');
-        setCurrentStagedId(stagedDatasetId);
-        datasetImportDialogState.open();
-    };
-
-    return (
-        <View
-            position='relative'
-            borderColor='gray-200'
-            borderRadius='regular'
-            backgroundColor='gray-75'
-            borderWidth='thin'
-        >
-            <View padding='size-150'>
-                <Flex justifyContent='space-between' alignItems='center' gap='size-250'>
-                    <Text UNSAFE_style={{ fontWeight: 500, fontSize: dimensionValue('size-200') }}>
-                        Import dataset - {fileName} - {formatBytes(stagedDataset?.size ?? 0)}
-                    </Text>
-
-                    <Divider size='S' marginY='size-150' />
-
-                    <Flex justifyContent='space-between' alignItems='center' gap='size-250'>
-                        <DeleteStagedFileConfirmation
-                            stagedDatasetId={stagedDatasetId}
-                            deleteEntry={() => deleteImportEntry(stagedDatasetId)}
-                        />
-
-                        <Button
-                            aria-label='continue dataset import'
-                            onPress={handleOpen}
-                            isDisabled={isError || isFetching}
-                        >
-                            Continue
-                        </Button>
-                    </Flex>
-                </Flex>
-
-                <Divider size='S' marginY='size-150' />
-
-                <Flex alignItems='center' gap='size-100'>
-                    <InfoOutline width={16} height={16} />
-
-                    <Text>{isError ? `Error: ${error?.detail}` : 'Map labels for the uploaded dataset'}</Text>
-                </Flex>
-            </View>
-        </View>
-    );
-};
