+++ application/ui/src/features/project/list/import-dataset-as-new-project/import-task-selection/import-task-selection.component.tsx
@@ -6,7 +6,7 @@ import { useActionState, useState } from 'react';
 import { Flex, Form, Item, Picker, Text, TextField, View } from '@geti/ui';
 import { InfoOutline } from '@geti/ui/icons';
 import { useProjects } from 'hooks/api/project.hook';
-import { useStagedDataset } from 'hooks/api/staged-dataset.hook';
+import { useStagedDatasetSuspense } from 'hooks/api/staged-dataset.hook';
 import { useImportDatasetAsNewProject } from 'hooks/localStorage/use-import-dataset-as-new-project.hook';
 
 import { TaskType } from '../../../../../constants/shared-types';
@@ -43,7 +43,7 @@ const useFormConfig = (stagedDatasetId: string, defaultTaskType: TaskType) => {
 
 export const ImportTaskSelection = ({ stagedDatasetId }: ImportTaskSelectionProps) => {
     const { data: projects } = useProjects();
-    const { data: stagedDataset } = useStagedDataset(stagedDatasetId);
+    const { data: stagedDataset } = useStagedDatasetSuspense(stagedDatasetId);
 
     const defaultTaskType = getRecommendedTaskType(stagedDataset?.metadata?.annotation_type);
     const [formState, submitAction] = useFormConfig(stagedDatasetId, defaultTaskType);
