+++ application/ui/src/hooks/api/staged-dataset.hook.ts
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { $api } from '../../api/client';
+import { isNonEmptyString } from '../../shared/util';
 import { isInvalidStagedFile } from './util';
 
 type useDeleteStagedDatasetProps = {
@@ -12,6 +13,17 @@ type useDeleteStagedDatasetProps = {
 };
 
 export const useStagedDataset = (stagedDatasetId: string | null | undefined) => {
+    return $api.useQuery(
+        'get',
+        '/api/staged_datasets/{staged_dataset_id}',
+        {
+            params: { path: { staged_dataset_id: stagedDatasetId } },
+        },
+        { enabled: isNonEmptyString(stagedDatasetId) }
+    );
+};
+
+export const useStagedDatasetSuspense = (stagedDatasetId: string | null | undefined) => {
     return $api.useSuspenseQuery('get', '/api/staged_datasets/{staged_dataset_id}', {
         params: { path: { staged_dataset_id: stagedDatasetId } },
     });
